refactor(MOR-2059): move the layout-mode vocabulary into presentation/
The v3 acceptance criterion requires a skin author to build a skin
without touching anything below the presentation layer. Until now they
had to edit lib/stores/layout.svelte.ts for their skin to be
selectable, and that same file was on the skins' forbidden-import
list — required and forbidden at once.

Move LayoutMode, CanonicalLayoutMode, CANONICAL_LAYOUT_MODES,
LEGACY_LAYOUT_ALIASES and normalizeLayoutMode verbatim into a new pure
module, presentation/layout-mode.ts: no localStorage, no store import,
no module-scope side effect. lib/stores/layout.svelte.ts becomes a
re-export shim — it cannot be deleted, since
lib/runtime/adapters/layout-mode-adapter.ts is eslint-banned from
importing presentation/** directly and reaches the vocabulary only
through this shim. getLayoutMode/setLayoutMode stay in the shim; they
are workspace-store-coupled, not vocabulary.

Point presentation/workspace/contract.ts at the new module instead of
hand-pinning its own copy of the layout-mode id space, and drop the
two contract.test.ts assertions that scraped lib/stores/layout.svelte.ts's
source text to keep that pinned copy in sync — a real import makes the
scrape redundant. The alias-count pin (exactly 4 known aliases) is kept,
now sourced from the real import rather than deleted outright, since
that fact was not otherwise covered.

Extend the three workspace-purity closure pins
(purity/store-purity/legacy-readers-purity.isolated.test.ts) to include
the new module in their asserted transitive closures, following the
same idiom PR #2883 used for the layout-compatibility guard. Their
titles reflect the delta only; the two pre-existing undercounts noted
in the ticket are left as they were.

Delete cycleLayoutMode (owner-approved 2026-08-31): zero production
callers, and not among the four symbols
docs/architecture/building-a-skin.md documents as the skin-authoring
surface. Its two remaining test callers (lib/stores/__tests__/layout.test.ts,
presentation/workspace/__tests__/selection-adoption.test.ts) are updated
in this same commit since the shim's removal of the function and their
removal of its calls cannot be split without an intermediate broken
state.
