+++ frontend/src/presentation/workspace/__tests__/purity.isolated.test.ts
@@ -96,13 +96,16 @@ describe('the workspace contract is pure at load (MOR-1077)', () => {
   });
 
   // ── Pin 2: structural, transitive over the enumerated closure ───────────
-  it('the closure is exactly the contract plus the two pure presentation contracts and the layout-compatibility guard', () => {
+  it('the closure is exactly the contract plus the two pure presentation contracts, the layout-compatibility guard and the layout-mode module', () => {
     expect(closure(ENTRY).sort()).toEqual([
       'src/presentation/languages/contract.ts',
       // MOR-2054: dev-only, pass-through guard `contract.ts` calls from
       // `registerDesignLanguage` — imports nothing but `contract.ts` itself
       // (type-only), so it is pure and belongs in this closure too.
       'src/presentation/languages/layout-compatibility-guard.ts',
+      // MOR-2059: the pure layout-mode vocabulary this contract now imports
+      // instead of hand-pinning its own copy. Imports nothing itself.
+      'src/presentation/layout-mode.ts',
       'src/presentation/layouts/contract.ts',
       ENTRY,
     ].sort());
