+++ frontend/src/presentation/workspace/__tests__/store-purity.isolated.test.ts
@@ -101,13 +101,16 @@ describe('the workspace store is inert until it is initialised (MOR-1079)', () =
     expect([...keys].sort()).toEqual(['rigplane:workspace', 'rigplane:workspace-migrated:v1']);
   });
 
-  it('the closure is the store, the repository, the three pure contracts and the layout-compatibility guard', () => {
+  it('the closure is the store, the repository, the three pure contracts, the layout-compatibility guard and the layout-mode module', () => {
     expect(closure(ENTRY).sort()).toEqual([
       'src/presentation/languages/contract.ts',
       // MOR-2054: dev-only, pass-through guard `contract.ts` calls from
       // `registerDesignLanguage` — imports nothing but `contract.ts` itself
       // (type-only), so it is pure and belongs in this closure too.
       'src/presentation/languages/layout-compatibility-guard.ts',
+      // MOR-2059: the pure layout-mode vocabulary `workspace/contract.ts`
+      // now imports instead of hand-pinning its own copy. Imports nothing.
+      'src/presentation/layout-mode.ts',
       'src/presentation/layouts/contract.ts',
       'src/presentation/workspace/contract.ts',
       'src/presentation/workspace/legacy-readers.ts',
